test: 테스트용 도메인 생성 팩토리 추가
