+++ src/test/java/kitchenpos/acceptance/MenuGroupAcceptanceTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.acceptance;
 
+import static kitchenpos.fixture.DomainFactory.createMenuGroup;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.RestAssured;
@@ -14,7 +15,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.jdbc.Sql;
 
 @DisplayName("메뉴 그룹 관련 기능")
 public class MenuGroupAcceptanceTest extends AcceptanceTest {
@@ -94,13 +94,5 @@ public class MenuGroupAcceptanceTest extends AcceptanceTest {
                 .collect(Collectors.toList());
 
         assertThat(resultMenuGroupIds).containsAll(expectedMenuGroupIds);
-    }
-
-    public static MenuGroup createMenuGroup(Long id, String name) {
-        MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(id);
-        menuGroup.setName(name);
-
-        return menuGroup;
     }
 }
